Support marking packages as hidden via command-line.
This adds support for marking specific packages as hidden via
command-line switch hidePackage. This is useful when the package in
question cannot be marked with @hide, for example, because the
package name in .class file is different from package name in source
file, or because the package may need to be visible in another body
of documentation.

Bug: 15073623
Change-Id: Iffcf5f3d50cc8db25a3bbfb244ce74c287e3875f