Change Doclava to support multimodule samples
Doclava only recognizes samples as browseable projects if they contain both
an _index.jd file and a src/ directory on their root directories. This CL
relaxes this rule by allowing the src/ to be inside a top level director as
well, so samples with multiple modules, like Wear apps, can be browsed.

Bug: 17042843

Change-Id: I2bcebb1dc3b2d317b43ff4af379c761067535bd9